refactor: move format_yyyy_mm_dd to general folder
